v4.23.0: New Command: Signup; Bug Fixes; Ditched Music Bot.

+++ src/main/java/useful/reminder/ReminderCommand.java
@@ -54,7 +54,7 @@ public class ReminderCommand extends Command {
             return;
         }
 
-        event.getMessage().delete().queue();
+        if (event.getGuild() != null) event.getMessage().delete().queue();
 
         var argsString = event.getArgs();
         var args = argsString.split(" ");
@@ -98,7 +98,7 @@ public class ReminderCommand extends Command {
         // Statistics.
         try {
             new User(event.getAuthor().getIdLong()).incrementFeatureCount(name.toLowerCase());
-            new Guild(event.getGuild().getIdLong()).incrementFeatureCount(name.toLowerCase());
+            if (event.getGuild() != null) new Guild(event.getGuild().getIdLong()).incrementFeatureCount(name.toLowerCase());
         } catch (SQLException e) {
             new Log(e, event.getGuild(), event.getAuthor(), name, event).sendLog(false);
         }
